+++ src/components/mentors/courses.tsx
@@ -13,7 +13,8 @@ import Pagination from "../re-usable/pagination";
 import { ToastContainer, toast, Slide } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Swal from "sweetalert2";
-import Cookies from "js-cookie";
+import { MENTOR_SERVICE_URL } from "@/utils/constant";
+import axios from "axios";
 
 interface CoursesData {
   _id?: string;
@@ -50,9 +51,16 @@ const MentorCourses = () => {
         toast.warn(error.response.data.message);
         return
       }
+      if(error && error?.response?.status === 401 && error.response?.data?.message === 'Mentor Not Verified'){
+        toast.warn(error?.response?.data?.message);
+        setTimeout(() => {
+          window.location.replace('/pages/mentor/profile');
+        }, 3000);
+        return;
+      }
       if (error && error.response?.status === 401) {
         toast.warn(error.response.data.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/mentor/login');
@@ -65,7 +73,7 @@ const MentorCourses = () => {
         error?.response?.data?.message === 'Mentor Blocked'
       ) {
         toast.warn(error?.response?.data?.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/mentor/login');
@@ -138,7 +146,7 @@ const MentorCourses = () => {
       }
       if (error && error.response?.status === 401) {
         toast.warn(error.response.data.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/mentor/login');
@@ -151,7 +159,7 @@ const MentorCourses = () => {
         error?.response?.data?.message === 'Mentor Blocked'
       ) {
         toast.warn(error?.response?.data?.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/mentor/login');
@@ -183,7 +191,7 @@ const MentorCourses = () => {
     } catch (error: any) {
       if (error && error.response?.status === 401) {
         toast.warn(error.response.data.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/student/login');
@@ -196,7 +204,7 @@ const MentorCourses = () => {
         error?.response?.data?.message === 'Mentor Blocked'
       ) {
         toast.warn(error?.response?.data?.message);
-        Cookies.remove('accessToken');
+        await axios.post(`${MENTOR_SERVICE_URL}/mentor/logout`, {}, { withCredentials: true }); //mentor logout api
         localStorage.clear();
         setTimeout(() => {
           window.location.replace('/pages/mentor/login');
